chore(Path, etc.): move Path file into dto
