+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -3,7 +3,7 @@ package nextstep.subway.path.application;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.fare.domain.DistanceExtraFare;
 import nextstep.subway.fare.domain.Fare;
-import nextstep.subway.path.domain.Path;
+import nextstep.subway.path.dto.Path;
 import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.application.StationService;
