+++ src/main/java/nextstep/subway/path/dto/Path.java
@@ -1,5 +1,6 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.path.dto;
 
+import nextstep.subway.path.domain.SectionEdge;
 import nextstep.subway.station.domain.Station;
 
 import java.util.List;
