+++ src/test/java/nextstep/subway/path/PathFinderTest.java
@@ -1,7 +1,7 @@
 package nextstep.subway.path;
 
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.path.domain.Path;
+import nextstep.subway.path.dto.Path;
 import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
